Time: 0 ms (100%), Space: 10 MB (89.82%) - LeetHub

+++ 0198-house-robber/0198-house-robber.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int rob(vector<int>& nums) {
-        int n = nums.size();
+        short n = nums.size();
         if(n == 1) return nums[0];
         else if(n == 2) return max(nums[1] , nums[0]);
         int odd = nums[0];
